# traffic-safety-openapi/public/api/docs/openapi.json
openapi: 3.0.0
info:
  title: 'Traffic Safety Platform API'
  description: 'API documentation for the Traffic Safety decision management system.'
  version: 1.0.0
servers:
  -
    url: /
    description: 'Relative base URL'
paths:
...
  /api/users:
    get:
      tags:
        - Users
      summary: 'List users.'
      operationId: listUsers
      responses:
        '200':
          description: 'User collection.'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/User' } }
                type: object
      security:
        -
          bearerAuth: []
    post:
      tags:
        - Users
      summary: 'Register a new user.'
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
              type: object
              additionalProperties: true
      responses:
        '201':
          description: 'User created.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          description: 'Invalid payload.'
      security:
        -
          bearerAuth: []
  '/api/users/{id}':
    get:
      tags:
        - Users
      summary: 'Get a user by id.'
      operationId: getUser
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'User details.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: 'User not found.'
      security:
        -
          bearerAuth: []
    put:
      tags:
        - Users
      summary: 'Update user information.'
      operationId: updateUser
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: 'User updated.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: 'User not found.'
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - Users
      summary: 'Delete a user.'
      operationId: deleteUser
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 'User deleted.'
      security:
        -
          bearerAuth: []
  /api/users/by-email:
    get:
      tags:
        - Users
      summary: 'Find a user by email address.'
      operationId: getUserByEmail
      parameters:
        -
          name: email
          in: query
          required: true
          schema:
            type: string
            format: email
      responses:
        '200':
          description: 'User details.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: 'User not found.'
      security:
        -
          bearerAuth: []
  '/api/users/{id}/role':
    post:
      tags:
        - Users
      summary: "Change a user's role."
      operationId: changeUserRole
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - role
              properties:
                role:
                  type: string
              type: object
      responses:
        '200':
          description: 'Role updated.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          description: 'Invalid role.'
      security:
        -
          bearerAuth: []
  '/api/users/{id}/record-login':
    post:
      tags:
        - Users
      summary: 'Record a login timestamp for a user.'
      operationId: recordUserLogin
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: false
        content:
          application/json:
            schema:
              properties:
                loggedInAt:
                  type: string
                  format: date-time
                  nullable: true
              type: object
      responses:
        '204':
          description: 'Login recorded.'
      security:
        -
          bearerAuth: []
  '/api/users/{id}/activate':
    post:
      tags:
        - Users
      summary: 'Activate a user.'
      operationId: activateUser
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 'User activated.'
      security:
        -
          bearerAuth: []
  '/api/users/{id}/deactivate':
    post:
      tags:
        - Users
      summary: 'Deactivate a user.'
      operationId: deactivateUser
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 'User deactivated.'
      security:
        -
          bearerAuth: []
components:
  schemas:
...
    User:
      required:
        - id
        - email
        - firstName
        - lastName
        - role
        - isActive
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        displayName:
          type: string
        role:
          type: string
          example: analyst
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      bearerFormat: JWT
      scheme: bearer
...